feat: add messageUpdate handler in case embed doesn't load immediately
